requirements.txt
Django==5.0.3
pytest-django==4.8.0
pytest-factoryboy==2.7.0
pytest-mock==3.12.0
djangorestframework==3.14.0
pre-commit==3.6.2
